+++ apps/remnant2toolkit/app/_components/genesis-code.tsx
@@ -11,6 +11,7 @@ function GlyphImage({ digit }: { digit: string }) {
       width={36}
       height={36}
       alt={digit}
+      quality={74}
     />
   );
 }
@@ -27,13 +28,15 @@ export function GenesisCode() {
       <div className="flex w-full flex-col items-center justify-center">
         <div
           id="code-container"
-          className="grid grid-cols-2 gap-y-4 sm:gap-x-12 sm:gap-y-0"
+          className="grid grid-cols-2 gap-y-4 bg-black px-4 dark:bg-transparent sm:gap-x-12 sm:gap-y-0"
         >
           <div
             id="code-top-row"
             className="col-span-full flex flex-col items-center justify-center sm:col-span-1"
           >
-            <h3 className="text-md mb-1 font-bold underline">Top Row</h3>
+            <h3 className="text-md mb-1 font-bold text-white underline">
+              Top Row
+            </h3>
             <div className="flex flex-row items-center justify-center gap-x-2">
               {topCode.split('').map((digit, index) => (
                 <GlyphImage key={`top-row-${index}`} digit={digit} />
@@ -44,7 +47,9 @@ export function GenesisCode() {
             id="code-bottom-row"
             className="col-span-full flex flex-col items-center justify-center sm:col-span-1"
           >
-            <h3 className="text-md mb-1 font-bold underline">Bottom Row</h3>
+            <h3 className="text-md mb-1 font-bold text-white underline">
+              Bottom Row
+            </h3>
             <div className="flex flex-row items-center justify-center gap-x-2">
               {bottomCode.split('').map((digit, index) => (
                 <GlyphImage key={`bottom-row-${index}`} digit={digit} />